(vanilla_csv_writer): Refactoring and managed to create and write to csv file
