+++ writer/src/test/resources/output.csv
@@ -1,0 +1,3 @@
+
+100,500,foo,bar
+500,3000,baz,boom
